MaP content updates
Updated the not guilty pleas for both defendant and company rep to

Pleading not guilty to this charge means you do not come to court on
the hearing date in your requisition pack - we'll send details of a new
hearing date.

If you disagree with any evidence from a witness statement in the
requisition pack – tell us the name of the witness and what you
disagree with the name, address and date of birth of any witnesses you
want to call to support your case.